[SPARK-47377][PYTHON][CONNECT][TESTS][FOLLOWUP] Factor out more tests from `SparkConnectSQLTestCase`
### What changes were proposed in this pull request?
It is a fallowup of https://github.com/apache/spark/pull/45497 , and factor out more tests from `SparkConnectSQLTestCase`

It should also be the laster PR for this test :)

### Why are the changes needed?
for testing parallelism

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #45533 from zhengruifeng/continue_break_basic.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
